Smart Car with a Arduino Mega 2560 and a Gertduino
The Arduino Mega 2560 is connected to six ultrasonic sensors, three in
front and three in back.  Also connected to a bluetooth module and acts
as an i2c master.  The Gertduino sitting on top of a Raspberry Pi acts
as an i2c slave and is connected to a L293D motor shield which in turn
is connected to four motors of the car.